src/config.py
""" Configuration loader for environment variables and constants. """

import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    DB_SERVER = os.getenv("DB_SERVER")
    DB_DATABASE = os.getenv("DB_DATABASE")
    DB_USER = os.getenv("DB_USER")
    DB_PASSWORD = os.getenv("DB_PASSWORD")
    GEMINI_API_KEY = os.getenv("GEMINI_API_KEY")
