requirements.txt
pymssql
google-generativeai
chromadb
sentence-transformers
streamlit
python-dotenv
pyodbc
